Misc. rendering improvements, added support for borders

+++ src/main/java/myworld/obsidian/geometry/Point2D.java
@@ -16,5 +16,5 @@
 
 package myworld.obsidian.geometry;
 
-public record Point2D(double x, double y) {
+public record Point2D(float x, float y) {
 }
